+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
   - Fix issue when using skipCodeCoverage=true
   - Check if a deployment is only caused by coverage failure but components and tests are ok (testCoverageNotBlocking)
   - Deploy with NoTestRun if deployment simulation has been performed with testCoverageNotBlocking=true
+- [hardis:project:configure:auth](https://sfdx-hardis.cloudity.com/hardis/project/configure/auth/): Fix creation of External Client App
 
 ## [6.22.0] 2025-01-13