+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 - [hardis:project:configure:auth](https://sfdx-hardis.cloudity.com/hardis/project/configure/auth/) enhancements & fixes:
   - Fix creation of External Client App
   - Improve prompts when selecting branches (filter those with / in their name)
+- Update default api version to 64.0
 
 ## [6.22.0] 2025-01-13