+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
 - [hardis:project:deploy:smart](https://sfdx-hardis.cloudity.com/hardis/project/deploy/smart/) enhancements & fixes:
+  - Implement usage of **RunRelevantTests** deployment mode
   - Fix issue when using skipCodeCoverage=true
   - Check if a deployment is only caused by coverage failure but components and tests are ok (testCoverageNotBlocking)
   - Deploy with NoTestRun if deployment simulation has been performed with testCoverageNotBlocking=true