+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
   - Fix issue when using skipCodeCoverage=true
   - Check if a deployment is only caused by coverage failure but components and tests are ok (testCoverageNotBlocking)
   - Deploy with NoTestRun if deployment simulation has been performed with testCoverageNotBlocking=true
+  - smartDeploymentTests: Add more metadata types that can not break apex test classes
 - [hardis:project:configure:auth](https://sfdx-hardis.cloudity.com/hardis/project/configure/auth/) enhancements & fixes:
   - Fix creation of External Client App
   - Improve prompts when selecting branches (filter those with / in their name)